Fix AllocateArrayBuffer By BoundFunction NewTarget
This is a historical problem:At first, we didn't support the full feature of newTarget in ECMA spec，in order to pass some cases of test262, hack the deleted code in AllocateArrayBuffer. In fact, the object so allocated is not an arrayBuffer.

Now, we support the full feature of newTarget, just delete it and do as ECMA spec.

add allocatearraybuffer unit test case.

Signed-off-by: ding <dingding5@huawei.com>
Change-Id: I2da7d8ce3dd58a63454ac9ecea4320458323fab9